requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.31.0

# For testing
coverage>=7.0.0
pytest>=7.4.0
pytest-django>=4.5.0

# Code quality
black>=23.0.0
flake8>=6.0.0
